pyproject.toml
[tool.poetry]
name = "review-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
openai = "^1.52.2"
langgraph = "^0.2.39"
fastapi = "^0.115.3"
pandas = "^2.2.3"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
qdrant-client = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
